reverted to using ints for solution to q1 as I think it may cause errors in some circumstances

+++ src/main/java/answers/Question1.java
@@ -8,8 +8,8 @@ import answers.TwoBitSplit;
 
 public class Question1 {
 	//Constants that define both the problem and the way I solve it
-	public static byte NUMBER_OF_BITS = 16;
-	public static byte BITS_OBSERVED = 1;
+	public static int NUMBER_OF_BITS = 16;
+	public static int BITS_OBSERVED = 1;
 	
 	public static int bestMergedPortfolio(int[] portfolios) {
 		//Edge case: if there is no portfolios or 1 portfolio "combining" 2 portfolios has no meaning so return 0
@@ -17,12 +17,12 @@ public class Question1 {
 			return 0;
 		}
 		//Initialise shift
-		byte shift = (byte)(NUMBER_OF_BITS-BITS_OBSERVED);
-		byte depth=0;
+		int shift = (int)(NUMBER_OF_BITS-BITS_OBSERVED);
+		int depth=0;
 		
 		//Splits array into 1 and 0 as we know that these will be a match since index < 100 we can store as a byte
-		ArrayList<Byte> M_0 = new ArrayList<Byte>(); //0
-		ArrayList<Byte> M_1 = new ArrayList<Byte>(); //1
+		ArrayList<Integer> M_0 = new ArrayList<Integer>(); //0
+		ArrayList<Integer> M_1 = new ArrayList<Integer>(); //1
 
 		ArrayList<Match> bestMatches = new ArrayList<Match>();
 		boolean started = false;
@@ -33,7 +33,7 @@ public class Question1 {
 		//HENCE THIS SOLUTION IS INVALID IF N>128
 		
 		while(!started) {
-			for(byte i=0; i<portfolios.length; i++) {
+			for(int i=0; i<portfolios.length; i++) {
 				//In case the input has incorrectly contained signed ints
 				//If done here no need to do it elsewhere as all signed ints
 				//will be ignored in further steps
@@ -49,7 +49,7 @@ public class Question1 {
 				//No matches found for the digit hence all must be 1 or all must 0 (unlikely) but a possibility
 				if(depth<=15) {
 					depth ++;
-					shift = (byte) (NUMBER_OF_BITS - (BITS_OBSERVED*(depth+1)));
+					shift = (int) (NUMBER_OF_BITS - (BITS_OBSERVED*(depth+1)));
 					M_0.clear(); M_1.clear();
 				} else {
 					//Went all the way down the chain and found 0 matches
@@ -63,7 +63,7 @@ public class Question1 {
 		int startDepth = depth;
 		
 		//Initialise the first nodes with 2 sets which are known to have matched
-		TwoBitSplit child = new TwoBitSplit((byte) (depth+1));
+		TwoBitSplit child = new TwoBitSplit((int) (depth+1));
 
 		//Calculate which child (or both) has longest consecutive streak
 		depth = child.split(portfolios, M_0, M_1);
